feat(cart): implement CartContext  - - Moved shopping cart logic into CartContext. Still owkring in progress
